Rename `chain/context` -> `chain/index`
